refactor(fomo-web): share stock swipe deck
